allow arguments for judgment task added new labels for the task
